Directory structure for different Zooming Algorithms
